+++ app/controllers/pdl.py
@@ -35,12 +35,15 @@ RCPP_STATE_DISTRIBUTION_DATA_JSON = "rcpp_state_distribution_data.json"
 RCPP_SUBPROGRAMS_DATA_JSON = "rcpp_subprograms_data.json"
 SNAP_JSON_DATA_PATH = os.path.join("controllers", "data", "snap")
 SNAP_DATA_JSON = "snap_state_distribution_data.json"
+SNAP_STATE_DISTRIBUTION_DATA_JSON = "snap_state_distribution_data.json"
+SNAP_SUBPROGRAMS_DATA_JSON = "snap_subprograms_data.json"
 EQIP_JSON_DATA_PATH = os.path.join("controllers", "data", "conservation", "eqip")
 EQIP_MAP_DATA_JSON = "eqip_map_data.json"
 EQIP_STATE_DISTRIBUTION_DATA_JSON = "eqip_state_distribution_data.json"
 EQIP_PRACTICE_CATEGORIES_DATA_JSON = "eqip_practice_categories_data.json"
 SNAP_JSON_DATA_PATH = os.path.join("controllers", "data", "snap")
 SNAP_DATA_JSON = "snap_state_distribution_data.json"
+SNAP_SUBPROGRAMS_DATA_JSON = "summary.json"
 CROP_INSURANCE_JSON_DATA_PATH = os.path.join("controllers", "data", "crop-insurance")
 CROP_INSURANCE_SUMMARY_DATA_JSON = "crop_insurance_subprograms_data.json"
 CROP_INSURANCE_STATE_DISTRIBUTION_DATA_JSON = "crop_insurance_state_distribution_data.json"
@@ -56,6 +59,12 @@ II_CSP_DATA_PATH = os.path.join(TITLE_II_DATA_PATH, "programs", "csp")
 II_CRP_DATA_PATH = os.path.join(TITLE_II_DATA_PATH, "programs", "crp")
 II_ACEP_DATA_PATH = os.path.join(TITLE_II_DATA_PATH, "programs", "acep")
 II_RCPP_DATA_PATH = os.path.join(TITLE_II_DATA_PATH, "programs", "rcpp")
+
+TITLE_IV_DATA_PATH = os.path.join("controllers", "data", "title-iv")
+IV_SNAP_DATA_PATH = os.path.join(TITLE_IV_DATA_PATH, "programs", "snap")
+
+TITLE_XI_DATA_PATH = os.path.join("controllers", "data", "title-xi")
+XI_CROP_INS_DATA_PATH = os.path.join(TITLE_XI_DATA_PATH, "programs", "crop-insurance")
 
 
 def search():
@@ -547,8 +556,8 @@ def titles_title_i_subtitles_subtitle_a_state_distribution_search():
     return data_json
 
 
-# /pdl/titles/title-i/subtitles/subtitle-a/subprograms:
-def titles_title_i_subtitles_subtitle_a_subprograms_search():
+# /pdl/titles/title-i/subtitles/subtitle-a/summary:
+def titles_title_i_subtitles_subtitle_a_summary_search():
     # set the file path
     subtitle_data = os.path.join(I_SUBTITLE_A_DATA_PATH, COMMOD_SUBPROGRAMS_DATA_JSON)
 
@@ -577,8 +586,8 @@ def titles_title_i_subtitles_subtitle_d_state_distribution_search():
     return data_json
 
 
-# /pdl/titles/title-i/subtitles/subtitle-d/subprograms:
-def titles_title_i_subtitles_subtitle_d_subprograms_search():
+# /pdl/titles/title-i/subtitles/subtitle-d/summary:
+def titles_title_i_subtitles_subtitle_d_summary_search():
     # set the file path
     subtitle_data = os.path.join(I_SUBTITLE_D_DATA_PATH, DMC_SUBPROGRAMS_DATA_JSON)
 
@@ -607,8 +616,8 @@ def titles_title_i_subtitles_subtitle_e_state_distribution_search():
     return data_json
 
 
-# /pdl/titles/title-i/subtitles/subtitle-e/subprograms:
-def titles_title_i_subtitles_subtitle_e_subprograms_search():
+# /pdl/titles/title-i/subtitles/subtitle-e/summary:
+def titles_title_i_subtitles_subtitle_e_summary_search():
     # set the file path
     subtitle_data = os.path.join(I_SUBTITLE_E_DATA_PATH, SADA_SUBPROGRAMS_DATA_JSON)
 
@@ -652,8 +661,8 @@ def titles_title_ii_programs_eqip_state_distribution_search():
         return data_json
 
 
-# /pdl/titles/title-ii/programs/eqip/practice-categories
-def titles_title_ii_programs_eqip_practice_categories_search():
+# /pdl/titles/title-ii/programs/eqip/summary
+def titles_title_ii_programs_eqip_summary_search():
     # set the file path
     eqip_data = os.path.join(II_EQIP_DATA_PATH, EQIP_PRACTICE_CATEGORIES_DATA_JSON)
 
@@ -697,8 +706,8 @@ def titles_title_ii_programs_csp_state_distribution_search():
         return data_json
 
 
-# /pdl/titles/title-ii/programs/csp/practice-categories
-def titles_title_ii_programs_csp_practice_categories_search():
+# /pdl/titles/title-ii/programs/csp/summary
+def titles_title_ii_programs_csp_summary_search():
     # set the file path
     csp_data = os.path.join(II_CSP_DATA_PATH, CSP_PRACTICE_CATEGORIES_DATA_JSON)
 
@@ -727,8 +736,8 @@ def titles_title_ii_programs_crp_state_distribution_search():
         return data_json
 
 
-# /pdl/titles/title-ii/programs/crp/subprograms
-def titles_title_ii_programs_crp_subprograms_search():
+# /pdl/titles/title-ii/programs/crp/summary
+def titles_title_ii_programs_crp_summary_search():
     # set the file path
     crp_data = os.path.join(II_CRP_DATA_PATH, CRP_SUBPROGRAMS_DATA_JSON)
 
@@ -757,8 +766,8 @@ def titles_title_ii_programs_acep_state_distribution_search():
         return data_json
 
 
-# /pdl/titles/title-ii/programs/acep/subprograms
-def titles_title_ii_programs_acep_subprograms_search():
+# /pdl/titles/title-ii/programs/acep/summary
+def titles_title_ii_programs_acep_summary_search():
     # set the file path
     acep_data = os.path.join(II_ACEP_DATA_PATH, ACEP_SUBPROGRAMS_DATA_JSON)
 
@@ -787,13 +796,73 @@ def titles_title_ii_programs_rcpp_state_distribution_search():
         return data_json
 
 
-# /pdl/titles/title-ii/programs/rcpp/subprograms
-def titles_title_ii_programs_rcpp_subprograms_search():
+# /pdl/titles/title-ii/programs/rcpp/summary
+def titles_title_ii_programs_rcpp_summary_search():
     # set the file path
     rcpp_data = os.path.join(II_RCPP_DATA_PATH, RCPP_SUBPROGRAMS_DATA_JSON)
 
     # open file
     with open(rcpp_data, 'r') as map_data:
+        file_data = map_data.read()
+
+        # parse file
+        data_json = json.loads(file_data, object_pairs_hook=OrderedDict)
+
+        return data_json
+
+
+# /pdl/titles/title-xi/programs/crop-insurance/state-distribution
+def titles_title_xi_programs_crop_insurance_state_distribution_search():
+    # set the file path
+    crop_ins_data = os.path.join(XI_CROP_INS_DATA_PATH, CROP_INSURANCE_STATE_DISTRIBUTION_DATA_JSON)
+
+    # open file
+    with open(crop_ins_data, 'r') as map_data:
+        file_data = map_data.read()
+
+        # parse file
+        data_json = json.loads(file_data, object_pairs_hook=OrderedDict)
+
+        return data_json
+
+
+# /pdl/titles/title-ii/programs/crop-insurance/summary
+def titles_title_xi_programs_crop_insurance_summary_search():
+    # set the file path
+    crop_ins_data = os.path.join(XI_CROP_INS_DATA_PATH, CROP_INSURANCE_SUMMARY_DATA_JSON)
+
+    # open file
+    with open(crop_ins_data, 'r') as map_data:
+        file_data = map_data.read()
+
+        # parse file
+        data_json = json.loads(file_data, object_pairs_hook=OrderedDict)
+
+        return data_json
+
+
+# /pdl/titles/title-iv/programs/snap/state-distribution
+def titles_title_iv_programs_snap_state_distribution_search():
+    # set the file path
+    snap_data = os.path.join(IV_SNAP_DATA_PATH, SNAP_STATE_DISTRIBUTION_DATA_JSON)
+
+    # open file
+    with open(snap_data, 'r') as map_data:
+        file_data = map_data.read()
+
+        # parse file
+        data_json = json.loads(file_data, object_pairs_hook=OrderedDict)
+
+        return data_json
+
+
+# /pdl/titles/title-iv/programs/snap/summary
+def titles_title_iv_programs_snap_summary_search():
+    # set the file path
+    snap_data = os.path.join(IV_SNAP_DATA_PATH, SNAP_SUBPROGRAMS_DATA_JSON)
+
+    # open file
+    with open(snap_data, 'r') as map_data:
         file_data = map_data.read()
 
         # parse file